requirements.txt
streamlit
google-generativeai
pydub
SpeechRecognition
gTTS
python-docx
python-dotenv

-e .